move FoodItemFilter to Filters namespace, not a db entity
